refactor: codingzone 도메인 관련 API 파일 이동
