+++ backend/src/main/java/com/icehufs/icebreaker/domain/codingzone/service/implement/CodingZoneServiceImplement.java
@@ -1,4 +1,4 @@
-package com.icehufs.icebreaker.service.implement;
+package com.icehufs.icebreaker.domain.codingzone.service.implement;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -12,24 +12,44 @@ import java.time.format.DateTimeFormatter;
 import java.time.temporal.TemporalAdjusters;
 import java.util.ArrayList;
 
-import com.icehufs.icebreaker.dto.object.CodingZoneStudentListItem;
-import com.icehufs.icebreaker.dto.object.PersAttendManagListItem;
-import com.icehufs.icebreaker.dto.object.ReservedClassListItem;
-import com.icehufs.icebreaker.dto.request.codingzone.*;
+import com.icehufs.icebreaker.domain.codingzone.dto.request.CodingZoneClassAssignRequestDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.request.GroupInfUpdateRequestDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.request.HandleAuthRequestDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.request.PatchGroupInfRequestDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.AuthorityExistResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.CodingZoneCanceResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.CodingZoneClassAssignResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.CodingZoneRegisterResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.DeleteAllInfResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.DeleteClassResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.DepriveAuthResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.GetCodingZoneAssitantListResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.GetCodingZoneStudentListResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.GetCountOfAttendResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.GetListOfCodingZoneClassForNotLogInResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.GetListOfCodingZoneClassResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.GetListOfGroupInfResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.GetPersAttendListItemResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.GetReservedClassListItemResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.GiveAuthResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.GroupInfUpdateResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.dto.response.PutAttendanceResponseDto;
+import com.icehufs.icebreaker.domain.codingzone.domain.vo.CodingZoneStudentListItem;
+import com.icehufs.icebreaker.domain.codingzone.domain.vo.PersAttendManagListItem;
+import com.icehufs.icebreaker.domain.codingzone.domain.vo.ReservedClassListItem;
 import com.icehufs.icebreaker.dto.response.ResponseDto;
 import com.icehufs.icebreaker.domain.article.dto.response.CheckOwnOfArticleResponseDto;
-import com.icehufs.icebreaker.dto.response.codingzone.*;
 import com.icehufs.icebreaker.entity.AuthorityEntity;
-import com.icehufs.icebreaker.entity.CodingZoneClass;
-import com.icehufs.icebreaker.entity.CodingZoneRegisterEntity;
-import com.icehufs.icebreaker.entity.GroupInfEntity;
+import com.icehufs.icebreaker.domain.codingzone.domain.entity.CodingZoneClass;
+import com.icehufs.icebreaker.domain.codingzone.domain.entity.CodingZoneRegisterEntity;
+import com.icehufs.icebreaker.domain.codingzone.domain.entity.GroupInfEntity;
 import com.icehufs.icebreaker.entity.User;
 import com.icehufs.icebreaker.repository.AuthorityRepository;
-import com.icehufs.icebreaker.repository.CodingZoneClassRepository;
-import com.icehufs.icebreaker.repository.CodingZoneRegisterRepository;
+import com.icehufs.icebreaker.domain.repository.CodingZoneClassRepository;
+import com.icehufs.icebreaker.domain.repository.CodingZoneRegisterRepository;
 import com.icehufs.icebreaker.repository.GroupInfRepository;
 import com.icehufs.icebreaker.repository.UserRepository;
-import com.icehufs.icebreaker.service.CodingZoneService;
+import com.icehufs.icebreaker.domain.codingzone.service.CodingZoneService;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
